Remove old C++ methods for efficiency estimation
